+++ Modules/HR/Database/Seeders/HRDatabaseSeeder.php
@@ -24,6 +24,6 @@ class HRDatabaseSeeder extends Seeder
         $this->call(HrApplicantsTableSeeder::class);
         $this->call(HrApplicationRoundTableSeeder::class);
         $this->call(HrApplicationsTableSeeder::class);
-        $this->call(HRJobsSeeder::class);
+        $this->call(HRJobSeederTableSeeder::class);
     }
 }
